First implementation of Quoridor for two players
This is the original version of the Quoridor program posted to Slack on 9/17/2017.
It is a console application designed for two human players. It uses the keyboard for the moves.
A number moves your piece to the numbered square. x passes, h places a horizontal wall with its left bound at the given cooridinates, and v places a vertical wall with its top bound at the given cooridinates.